plant-species-classification/requirements.txt
# CT7160NI Computer Vision Coursework
# Plant Species Classification Project
# Python 3.8+ required

# ============================================
# Core Deep Learning Frameworks
# ============================================
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0                     # Pre-trained models (EfficientNet, etc.)

# ============================================
# Computer Vision Libraries
# ============================================
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0           # Advanced data augmentation

# ============================================
# Data Science & Machine Learning
# ============================================
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ============================================
# Visualization
# ============================================
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================
# Model Interpretability
# ============================================
grad-cam>=1.4.8
captum>=0.6.0                   # PyTorch interpretability library

# ============================================
# Experiment Tracking
# ============================================
tensorboard>=2.13.0

# ============================================
# Utilities
# ============================================
tqdm>=4.65.0                    # Progress bars
PyYAML>=6.0                     # Configuration management
jupyter>=1.0.0                  # Jupyter notebooks
ipywidgets>=8.0.0               # Notebook widgets

# ============================================
# Optional: Weights & Biases (Experiment Tracking)
# ============================================
# wandb>=0.15.0

